% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AlpineInt}
\alias{AlpineInt}
\title{Alpine absolute abundance data.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Early}{relative abundance of 45 alpine species sampled in early stages}
  \item{Mid}{relative abundance of 45 alpine species sampled in mid stages}
   \item{Late}{relative abundance of 45 alpine species sampled in late stages}
}}
\usage{
AlpineInt
}
\description{
A dataset containing the absolute abundance of 45 alpine species.
Based on the age of the glacial deposits, data were assigned to three
successional stages(columns)
}
\keyword{datasets}
