% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{dis1}
\alias{dis1}
\title{\code{dis1}: computes the contribution of each species/node to the two types of dissimilarity measures.}
\usage{
dis1(x, q, type = "tax", type2 = "species", tree = NULL)
}
\arguments{
\item{x}{the species-by-assemblages abundance matrix or data.frame with species names as rownames and the colnames as assemblage/site names.}

\item{q}{a single value for the diversity order.}

\item{type}{a string: "tax" (taxonomic) or "phy" (phylogenetic).}

\item{type2}{a string: "species" or "k"."species" means the contribution of each species/node to the two types of dissimilarity measures
(Jaccard-type dissimilarity and Sorensen-type dissimilarity).
"k" means the contribution of each assemblage/location/site to the two types of dissimilarity measures
(Jaccard-type dissimilarity and Sorensen-type dissimilarity). In the worked example, the contribution of each assemblage/stage is not computed.}

\item{tree}{the pylo object of the phylogenetic tree of all assemblages. Needed only if \code{type="phy"}.}
}
\value{
a data frame tabulating the contribution of each species/node to the two types of dissimilarity measures: Jaccard-type (1-U_qN) and Sorensen-type (1-C_qN)
}
\description{
\code{dis1}: computes the contribution of each species/node to the two types of dissimilarity measures.
}
\examples{
#Taxonomic analysis example
data(Alpine)
out0 <- dis1(x = Alpine, q = 0, type = "tax")
#Phylogenetic analysis example
data(Alpine)
data(tree_Alpine)
out0 <- dis1(x = Alpine, q = 0, type = "phy", tree = tree_Alpine)
}
