% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{draw_dis_spe}
\alias{draw_dis_spe}
\title{\code{draw_dis_spe} plots the contribution of each species/node to dissimilarity (Jaccard-type dissimilarity and Sorensen-type dissimilarity).}
\usage{
draw_dis_spe(x, title_name, type = "tax")
}
\arguments{
\item{x}{a merged table of output values with three columns corresponding to output for q = 0, 1, 2. See example for details.}

\item{title_name}{the title name of plot (Usually the Jaccard-type dissimilarity or Sorensen-type dissimilarity). See example for details.}

\item{type}{a string specifying the type of contribution: "tax" for taxonomic and "phy" for phylogenetic}
}
\value{
a ggplot2 object showing the contribution of each species/node.
}
\description{
\code{draw_dis_spe} plots the contribution of each species/node to dissimilarity (Jaccard-type dissimilarity and Sorensen-type dissimilarity).
}
\examples{
#Taxonomic analysis example
data(Alpine)
out0 <- dis1(x = Alpine, q = 0, type = "tax")
out1 <- dis1(x = Alpine, q = 1, type = "tax")
out2 <- dis1(x = Alpine, q = 0, type = "tax")
tax_UqN_r <- cbind(out0[, 1], out1[, 1], out2[, 1])
tax_CqN_r <- cbind(out0[, 2], out1[, 2], out2[, 2])
draw_dis_spe(tax_UqN_r, "Jaccard-type taxonomic dissimilarity")
draw_dis_spe(tax_CqN_r, "Sorensen-type taxonomic dissimilarity")
#Phylogenetic analysis example
data(Alpine)
data(tree_Alpine)
out0 <- dis1(x = Alpine, q = 0, type = "phy", tree = tree_Alpine)
out1 <- dis1(x = Alpine, q = 1, type = "phy", tree = tree_Alpine)
out2 <- dis1(x = Alpine, q = 2, type = "phy", tree = tree_Alpine)
phy_UqN_r <- cbind(out0[, 1], out1[, 1], out2[, 1])
phy_CqN_r <- cbind(out0[, 2], out1[, 2], out2[, 2])
draw_dis_spe(phy_UqN_r, "Jaccard-type phylogenetic dissimilarity", type = "phy")
draw_dis_spe(phy_CqN_r, "Sorensen-type phylogenetic dissimilarity", type = "phy")
}
