% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Evenness}
\alias{Evenness}
\title{\code{Evenness}: compute six evenness indices of order q.}
\usage{
Evenness(data, q = seq(0, 2, 0.1))
}
\arguments{
\item{data}{a data.frame, matrix or list of observed species-by-assemblage frequency vector. For the case
of only one assemblage (site), dat should be a numeric vector of species frequency.}

\item{q}{a vector of diversity orders: user must specify a vector (default  is from 0 to 2 in an increment
of 0.1)}
}
\value{
a list of N dataframes, where N is the number of assemblages (sites). Each data.frame shows
the profiles of all six classes of evenness indices listed in Chao and Ricotta (2019) Ecology paper.
}
\description{
\code{Evenness}: compute six evenness indices of order q.
}
\examples{
data(Alpine)
out1 <- Evenness(data = Alpine)
}
