% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{gg_evenness}
\alias{gg_evenness}
\title{\code{gg_evenness}: use ggplot2 and ggpubr to plot the outcome of function \code{Evenness}.}
\usage{
gg_evenness(x)
}
\arguments{
\item{x}{a optput of \code{Evenness}}
}
\value{
return a ggplot2 object
}
\description{
\code{gg_evenness}: use ggplot2 and ggpubr to plot the outcome of function \code{Evenness}.
}
\examples{
data(Alpine)
out1 <- Evenness(data = Alpine)
gg_evenness(out1)
}
