% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_Alpine}
\alias{tree_Alpine}
\title{Phlogenetic tree for the Alpine data}
\format{An object of class \code{phylo} of length 5.}
\usage{
tree_Alpine
}
\description{
A phylo object for the example data "Alpine".
}
\keyword{datasets}
