% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{gg_evenness}
\alias{gg_evenness}
\title{\code{gg_evenness}: use ggplot2 and ggpubr to plot the outcome of function \code{Evenness}.}
\usage{
gg_evenness(x)
}
\arguments{
\item{x}{a optput of \code{Evenness}}
}
\value{
return a ggplot2 object
}
\description{
\code{gg_evenness}: use ggplot2 and ggpubr to plot the outcome of function \code{Evenness}.
}
\examples{
\donttest{
data(AlpineInt)
out2 <- Evenness(data = AlpineInt,q = seq(0, 2, 0.1), nboot = 50)
gg_evenness(out2)
}
}
